MAJ 24/09/2024

+++ 11-Les_strings/11_3-Travailler_avec_strings/P11_3_c-Dizaine_un_nombre/src/main/kotlin/main.kt
@@ -1,6 +1,5 @@
 fun main() {
-    val nombre = readln().toInt()
-
+    val nombre = readln()
     println(nombre.reversed().substring(1, 2).reversed())
 }
